requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==3.1.1         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-auth==1.11.2       # via google-api-python-client, google-auth-httplib2, google-auth-oauthlib
httplib2==0.17.0          # via google-api-python-client, google-auth-httplib2
idna==2.9                 # via requests
oauthlib==3.1.0           # via requests-oauthlib
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via requests-oauthlib
rsa==4.0                  # via google-auth
six==1.14.0               # via google-api-python-client, google-auth
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
